Move Scoreboard to Canvas and fix collector state
Prior to this the Scoreboard was using a text component that
rendered on top of everything else.  Now it is using a wordspace
canvas which will be properly occluded by worldspace objects.  The
Collector script is now using a bool to represent activity state,
as the prior bitwise State() operations were confusing.