Add the basic skeleton for ControlSets.  Broken.
This commit adds IControlSet, AbstractControlSet, and a pair of
ControlSet implementations.  The idea here is that instead of tying
Catchable item dispensation to specific keystrokes, it will be
implemented by a strategy.  This allows for solo play (timed items)
as well as pair play (keystroke or click activated).  It would also
potentially allow for a network player, should that ever be
desired.