Add IControlSet loading to GameManager
This commit makes the changes needed to load and activate
IControlSets from inside of the GameManager, as well as fixing the
IControlSet bindings in the Installer, and fleshing out the
PartnerControlSet.  This fleshing out moves the responsibility for
triggering Catchable dispensing from the GameContainer to the
PartnerControlSet.  I'd really love to remove the GameContainer,
but unsure how that would work, as it's inital purpose was actually
to make sure the GameManager got fired up correctly.  It should no
longer be required, but I need to check in a later commit.

This commit also contains a small, unassociated bugfix for
Collector.  Prior behavior was to activate the bin as soon as a
PresentList was added.  This was fine, until the bin lid management
was added, triggering a bin score before it was even registered.
In the long run, this behavior shouldn't matter, as we should stop
injecting lists on bin create.