Add Skeleton of scoreboard display monobehavior
This commit is focused on getting details of the game state to the
scoreboard for display to the user.  It adds a GameDescriptor
struct that is intended for easy display, and a few methods for
building and passing that descriptor to the Game and
AbstractRoundInspector classes.  Currently, the game is not
displaying anything, as the event listeners are not hooked up.