Replace EventSource with modified GameEventBus
This is a major commit, reworking the event bus system for the
entire game.  The hacky unity tutorial event system has been
replaced with a slightly modified version of GameEventBus, which
needs it's own github repo. In writing this, I also got rid of a
few outstanding annoyances with things such as ItemDispensation.