Fix immortal async/await spawning loop
The await loop used by the SoloControlSet was incorrectly
continuing to run after the end of the play mode.  This resulted
in Catchables being spawned into the scene even while in edit mode.

Solution was to not rely on the destructor, instead, using the
IDisposable interface and changing how the interfaces were bound
in.